Netcode: rename ClientError to DisconnectionReason. Change disconnected state to contain reason.
This makes easier to check if the client is actually connected. Currently there is no way to create an client without a ConnectToken or change the token after created, so there is never an idle state, the client is always trying to connect or failed to.

In the future this might change, since it may be interesting to create a client without a ConnectToken and add a function that receives one token to start the connection.