ChatDemo: simplify demo and remove logger app.
This example was getting to complicated so I tried to simplify, removing the logger app helped, we are using env logger now.

Now the client/server is stored at the app state, instead of Options for server and client. It makes clear when each one is set and it is easier to understand transitions between state.

When a server is hosted. the client does not generates a RenetClient, instead we bypass it directly, if a message is sent by the host it is directly accepted and broadcast to the others clients. This removes the need to send packets when the client is the host.