Rechannel: introduce the Channel trait.
The Channel trait has made a come back. This time we keep the ChannelConfig as an enum but making it return a Boxed dyn Channel. This removes some duplicated code and helps identify the similarities of what is executed in each channel. 

Moved some errors to the ChannelError. Now the channels should have an error state that uses this enum.