Rechannel: add error for sending message above packet budget.
When a reliable message is sent, and the message size is above the packet budget, the message would never be sent, and the channel would soon be out of sync. So a new error was created for this case. 

In the unreliable channel case, the message would also be dropped, but in this case, the channel would just keep working. We set an error in the channel to warn the user. We could just drop this message silently and warn the user with a log. Need to check if this would be more desired with real use cases.