Rechannel: use Bytes for send_message
Currently, when broadcasting a message in the server, we make a copy for each connected client. This can adds a lot memory usage when the numbers of connected clients is high. Using bytes instead, we don't need to make copies, because it uses a reference count to manage the memory.