Renetcode: fix using global sequence when sending first KeepAlive packet
This packet was not using the client sequence, and was mistakenly using the global sequence when the packet was generated. This was causing false duplicated packets, when one packet is generated its sequence could match with the global sequence used when the client connected, making that packet miss. 

This means that only one packet would be ever missed because of this bug. The test were updated to verify this case.