Rechannel: add max_message_size configuration
This gives a finer control on how the channel should behave. 

Security: this fixes an exploit over the block channel. Was possible to send an SliceMessage with a high value of num_slices, this would cause the channel to allocate a lot of memory causing Out of Memories errors. Now we check to see if the memory to be allocated is in the allowed range.