Rechannel: split channels into Sender and Receiver
Now server and clients can have different channels. The code was already divided having no meaning for having the send/receive channel being the same. 

Now you need to a configuration for your server channels and for the client channels, but you have way more control over them. No need have channels that the server/client would never use, or have the client limit be the same as the channels.

ChannelNetworkInfo was removed since we would need to refactor it for it to be usable. Since it was not really being used I decided to cut it and another time revisit it.