DemoRoyale: update channels configuration
This is a good showcase of having different channel to send/receive from. 

The server receives 2 reliable channels from the client and the server sends 1 reliable  and 1 unreliable. Before, you needed 3 channels for this to be possible, and some channels would be unusable by the client and the server.

Also we can split the enum for the channels and give them a more meaningful name. Before we had Channel::Reliable that was not really telling what was inside it, and was hard naming it better because the server and client was using the same channel. Now we can use ClientChannel::Input and ServerChannel::ServerMessages and the intend becomes clearer .