Rechannel: don't ack fragment packet on the first received fragment
Only ack the packet when all fragments are received.

The connection was considering the packet received on the first fragment received, this would cause sync errors when one fragment is missed, the other end would think the packet was received and would never resend the messages that were in that packet.

Added test to considers check this case, also updated the usage test to fragment packets and add rng to have better packet loss simulation.