Renet: remove redundancy for disconnect packets
Whenever we send more than 1 disconnect packet after the first one is received, it was generating packet duplication errors. This was really confusing for some users since it logged an error (4 times most time), but that was expected to happen. Remove for now, since if the disconnect is lost, the user will just timeout.

We should have a function in renetcode to generate unique disconnect packet if we want to have redundancy.